Class_05/search_agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.9",
    "tavily-python>=0.7.11",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
